[IMP] appointment_filter: refactored filtering logic and improve XML structure
