+++ appointment_filter/__manifest__.py
@@ -2,14 +2,11 @@
     'name': 'Website Appointment Filters Extension',
     'version': '1.0',
     'category': 'Website',
-    'depends': ['website_appointment','appointment_account_payment'],
+    'depends': ['website_appointment', 'appointment_account_payment'],
     'summary': 'Extend website appointments with additional filters for type, schedule, and payment',
-    'author': 'Mandani Tushar',
- 
     'data': [
         'views/website_appointment_templates.xml',
     ],
     'installable': True,
-    'application': False,
     'license': 'LGPL-3',
-    }
+}
