+++ appointment_filter/controllers/__init__.py
@@ -1,1 +1,1 @@
-from . import controllers
+from . import appointment
